requirements.txt
# Core (sempre necessário)
python-dotenv>=1.0.0
langchain>=0.3.0
chromadb>=0.5.0
pandas>=2.0.0

# LLM Providers (instalar conforme necessário)
groq>=0.4.0        # Se usar Groq